pyproject.toml
[tool.poetry]
name = "hyperion-bot"
version = "0.1.0"
description = ""
authors = ["Riley Flynn <riley@rileyflynn.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {version = "^1.9.1", extras=["dotenv"]}
requests = "^2.27.1"
"discord.py" = {git = "https://github.com/rapptz/discord.py.git"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
